Fixes various compiler warnings.
git-svn-id: http://svn.mythtv.org/svn/trunk@23719 7dbf422c-18fa-0310-86e9-fd20926502f2

+++ mythtv/libs/libmythdb/mythdbcon.cpp
@@ -258,7 +258,7 @@ void MDBManager::PurgeIdleConnections(void)
     while (it != m_pool.end())
     {
         totalConnections++;
-        if ((*it)->m_lastDBKick.secsTo(now) <= kPurgeTimeout)
+        if ((*it)->m_lastDBKick.secsTo(now) <= (int)kPurgeTimeout)
         {
             ++it;
             continue;
